+++ lib/src/widget/control_widget/zoom_control_widget.dart
@@ -1,5 +1,6 @@
-part of flutter_naver_map;
+part of "../../../flutter_naver_map.dart";
 
+/// 줌 레벨을 컨트롤할 수 있는 위젯입니다.
 class NaverMapZoomControlWidget extends StatelessWidget {
   final NaverMapController? mapController;
   final bool nightMode;
@@ -15,8 +16,10 @@ class NaverMapZoomControlWidget extends StatelessWidget {
   });
 
   Color get buttonColor => nightMode ? Colors.grey.shade900 : Colors.white;
-  Color get iconColor => nightMode ? Colors.grey.shade200 : Colors.grey.shade800;
-  Color get dividerColor => nightMode ? Colors.grey.shade700 : Colors.grey.shade200;
+  Color get iconColor =>
+      nightMode ? Colors.grey.shade200 : Colors.grey.shade800;
+  Color get dividerColor =>
+      nightMode ? Colors.grey.shade700 : Colors.grey.shade200;
 
   @override
   Widget build(BuildContext context) {
@@ -26,8 +29,7 @@ class NaverMapZoomControlWidget extends StatelessWidget {
         borderRadius: BorderRadius.circular(roundness),
         color: buttonColor,
         elevation: 1,
-        child: Column(mainAxisSize: MainAxisSize.min,
-            children: [
+        child: Column(mainAxisSize: MainAxisSize.min, children: [
           baseWidget(icon: Icons.add, action: zoomIn),
           _divider(),
           baseWidget(icon: Icons.remove, action: zoomOut),
@@ -58,10 +60,10 @@ class NaverMapZoomControlWidget extends StatelessWidget {
     if (mapController == null) return;
     final nowZoomLevel =
         await mapController!.getCameraPosition().then((cp) => cp.zoom);
-    final update =
-        NCameraUpdate.withParams(zoom: (nowZoomLevel.round() + delta).toDouble())
-          ..setReason(NCameraUpdateReason.control)
-          ..setAnimation(duration: const Duration(milliseconds: 300));
+    final update = NCameraUpdate.withParams(
+        zoom: (nowZoomLevel.round() + delta).toDouble())
+      ..setReason(NCameraUpdateReason.control)
+      ..setAnimation(duration: const Duration(milliseconds: 300));
     mapController?.updateCamera(update);
   }
 }